merged_project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==0.21.1
Jinja2==3.1.4
WeasyPrint==59.0
pydyf==0.7.0
PyMuPDF==1.24.1
requests==2.31.0
python-multipart==0.0.9
pydantic
pydantic-settings
email-validator
openai
httpx

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.9
